feat(refactor-dev): Update the environment and fix the top menu bug - [更新环境，修复顶部菜单bug]
