+++ package.json
@@ -45,11 +45,11 @@
     "@adp/hooks": "workspace:*",
     "@adp/materials": "workspace:*",
     "@adp/utils": "workspace:*",
-    "@alova/scene-vue": "^1.5.0",
-    "@arco-design/web-vue": "^2.55.2",
+    "@alova/scene-vue": "^1.6.0",
+    "@arco-design/web-vue": "^2.55.3",
     "@arco-plugins/vite-vue": "^1.4.5",
-    "@vueuse/core": "^10.9.0",
-    "axios": "^1.6.8",
+    "@vueuse/core": "^10.10.1",
+    "axios": "^1.7.2",
     "dayjs": "^1.11.11",
     "mozjpeg": "7.1.1",
     "pinia": "^2.1.7",
@@ -57,18 +57,18 @@
     "vue": "^3.4.27",
     "vue-draggable-plus": "^0.4.1",
     "vue-i18n": "^9.13.1",
-    "vue-router": "^4.3.2",
+    "vue-router": "^4.3.3",
     "vue-types": "^5.1.2"
   },
   "devDependencies": {
     "@iconify/vue": "^4.1.2",
     "@types/lodash-es": "^4.17.12",
-    "@types/node": "^20.12.12",
+    "@types/node": "^20.14.2",
     "@types/nprogress": "^0.2.3",
     "@types/qs": "^6.9.15",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
-    "@unocss/preset-legacy-compat": "^0.60.2",
+    "@unocss/preset-legacy-compat": "^0.60.4",
     "@vitejs/plugin-vue": "^4.6.2",
     "dotenv": "^16.4.5",
     "eslint": "^8.57.0",
@@ -86,8 +86,8 @@
     "qs": "^6.12.1",
     "rollup-plugin-visualizer": "^5.12.0",
     "typescript": "^5.4.5",
-    "unocss": "^0.60.2",
-    "vite": "^5.2.11",
+    "unocss": "^0.60.4",
+    "vite": "^5.2.13",
     "vite-plugin-compression": "^0.5.1",
     "vite-plugin-html": "^3.2.2",
     "vite-plugin-imagemin": "^0.6.1",
