--- src/layout/common/menu/components/base-menu/components/index.ts
@@ -1,4 +1,0 @@
-import SubMenuItem from './sub-menu-item/index.vue';
-import MenuItem from './menu-item/index.vue';
-
-export { MenuItem, SubMenuItem };
